requirement.txt
Package                   Version
------------------------- --------------
attrs                     24.2.0
beautifulsoup4            4.12.3
GeoAlchemy2               0.16.0
geographiclib             2.0
geopy                     2.4.1
gpxpy                     1.6.2
jsonschema-specifications 2024.10.1
jupyter                   1.1.1
jupyter_client            8.6.3
jupyter-console           6.6.3
jupyter_core              5.7.2
jupyter-events            0.10.0
jupyter-lsp               2.2.5
jupyter_server            2.14.2
jupyter_server_terminals  0.5.3
jupyterlab                4.2.5
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
jupyterlab_widgets        3.0.13
lazy_loader               0.4
numpy                     2.1.3
opencv-python             4.10.0.84
overrides                 7.7.0
packaging                 24.1
pandas                    2.2.3
pillow                    10.4.0
pip                       24.3.1
psycopg2-binary           2.9.10
pyproj                    3.7.0
pytesseract               0.3.13
python-dateutil           2.9.0.post0
python-json-logger        2.0.7
pytz                      2024.2
PyYAML                    6.0.2
shapely                   2.0.6
SQLAlchemy                2.0.36
urllib3                   2.2.3
